refactor AddDeck component to use MarkdownEditor
